update projects endpoints

+++ coati/web/api/project.py
@@ -11,19 +11,26 @@ from coati.core.models.project import Project, Column, ProjectMember
 from coati.core.models.ticket import Ticket, Attachment
 from coati.web.api.auth import AuthResource
 from coati.utils import send_new_member_email_async, save_notification
+from coati.web.api import errors as api_errors, current_app
+from coati.web.api.auth.utils import current_user
 
 
 class ProjectList(AuthResource):
-    def __init__(self):
-        super(ProjectList, self).__init__()
+    """
+    Project Resource List
+    """
 
     def get(self):
-        prj_mem = ProjectMember.objects(member=g.user.id)
+        """
+        Get the list of projects that the user belongs
+        :return: List of project resources
+        """
+        prj_mem = ProjectMember.get_by_member(current_user)
         projects = []
         for pm in prj_mem:
             if pm.project.active:
                 projects.append(pm.project.to_dict())
-            elif pm.project.owner.id == g.user.id:
+            elif pm.project.owner.id == current_user.id:
                 projects.append(pm.project.to_dict())
         return projects, 200
 
@@ -31,17 +38,14 @@ class ProjectList(AuthResource):
         """
         Create Project
         """
-        data = request.get_json(force=True, silent=True)
+        data = request.get_json(silent=True)
         if not data:
-            msg = "payload must be a valid json"
-            return jsonify({"error": msg}), 400
-        try:
-            user = User.objects.get(pk=g.user_id)
-        except DoesNotExist, e:
-            return jsonify({"error": 'owner user does not exist'}), 400
+            return api_errors.InvalidAPIUsage(
+                api_errors.INVALID_JSON_BODY_MSG
+            )
 
         prj = Project(name=data.get('name'),
-                      owner=user.to_dbref())
+                      owner=g.user.to_dbref())
         prj.active = data.get('active')
         prj.private = data.get('private')
         prj.prefix = data.get('prefix', data.get('name')[:3].upper())
@@ -73,27 +77,58 @@ class ProjectList(AuthResource):
                           verb='new_project',
                           data=prj.to_dict())
 
-        return prj.to_json(), 201
+        return prj.to_dict(), 201
 
 
 class ProjectInstance(AuthResource):
-    def __init__(self):
-        super(ProjectInstance, self).__init__()
+    """
+    Project Resource
+    """
 
     def get(self, project_pk):
-        prj = Project.objects.get(pk=project_pk).select_related(max_depth=2)
+        """
+        Get a Project Instance
+        :param project_pk: the ID of the project
+        :return: a Project Object
+        """
+        prj = Project.get_by_id(project_pk).select_related(max_depth=2)
+        if prj is None:
+            return api_errors.MissingResource(
+                api_errors.INVALID_PROJECT_MSG
+            )
+
         return prj.to_dict(), 200
 
     def put(self, project_pk):
-        project = Project.objects.get(pk=project_pk)
-        data = request.get_json(force=True, silent=True)
+        """
+        Update a Project Instance
+        :param project_pk:
+        :return:
+        """
+        project = Project.get_by_id(project_pk)
+
+        if not project:
+            return api_errors.MissingResource(
+                api_errors.INVALID_PROJECT_MSG
+            )
+
+        data = request.get_json(silent=True)
+
         if not data:
-            msg = "payload must be a valid json"
-            return jsonify({"error": msg}), 400
+            return api_errors.InvalidAPIUsage(
+                api_errors.INVALID_JSON_BODY_MSG
+            )
+
         project.active = data.get('active')
         project.description = data.get('description')
         project.name = data.get('name')
-        owner = User.objects.get(id=data.get('owner_id'))
+        owner = User.get_by_id(data.get('owner_id'))
+
+        if not owner:
+            return api_errors.InvalidAPIUsage(
+                api_errors.INVALID_USER_ID_MSG
+            )
+
         project.owner = owner.to_dbref()
         project.private = data.get('private')
         project.sprint_duration = data.get('sprint_duration')
@@ -106,27 +141,53 @@ class ProjectInstance(AuthResource):
                           verb='update_project',
                           data=project.to_dict())
 
-        return project.to_json(), 200
+        return project.to_dict(), 200
+
 
     def delete(self, project_pk):
-        project = Project.objects.get(pk=project_pk)
+        """
+        Delete a Project Instance
+        :param project_pk: Project ID
+        :return: nothing
+        """
+        project = Project.get_by_id(project_pk)
+        if not project:
+            return api_errors.MissingResource(
+                api_errors.INVALID_PROJECT_MSG
+            )
+
         project.delete()
-        return jsonify({}), 204
+        return {}, 204
 
 
 class ProjectColumns(AuthResource):
-    def __init__(self):
-        super(ProjectColumns, self).__init__()
+    """
+    Columns Resource List
+    """
 
     def get(self, project_pk):
-        return Column.objects(project=project_pk).order_by('order').to_json()
+        """
+        :param project_pk: Project ID
+        :return: List of columns
+        """
+        return Column.get_by_project(project_pk).to_json()
 
     def post(self, project_pk):
-        data = request.get_json(force=True, silent=True)
+        """
+        :param project_pk: Project ID
+        :return: Column Resource created
+        """
+        data = request.get_json(silent=True)
         if not data:
-            msg = "payload must be a valid json"
-            return jsonify({"error": msg}), 400
-        project = Project.objects.get(pk=project_pk)
+            return api_errors.InvalidAPIUsage(
+                api_errors.INVALID_JSON_BODY_MSG
+            )
+
+        project = Project.get_by_id(project_pk)
+        if not project:
+            return api_errors.MissingResource(
+                api_errors.INVALID_PROJECT_MSG
+            )
 
         col = Column()
         col.order = Column.objects.count()
@@ -138,7 +199,7 @@ class ProjectColumns(AuthResource):
 
         # Check if already exists one Done column
         if col.done_column:
-            columns = Column.objects(project=project)
+            columns = Column.get_by_project(project)
             for c in columns:
                 if c.done_column:
                     c.done_column = False
@@ -150,57 +211,85 @@ class ProjectColumns(AuthResource):
                           verb='new_column',
                           data=col.to_dict())
 
-        return col.to_json(), 200
+        return col.to_dict(), 200
 
 
 class ProjectColumn(AuthResource):
-    def __init__(self):
-        super(ProjectColumn, self).__init__()
+    """
+    Columns Resource
+    """
 
     def get(self, project_pk, column_pk):
-        return Column.objects.get(pk=column_pk).to_json()
+        """
+        Get Column Instance
+        :param project_pk: Project that belongs
+        :param column_pk: Id of the column
+        :return: Object Column
+        """
+
+        return Column.get_by_id(column_pk).to_dict()
 
     def put(self, project_pk, column_pk):
-        col = Column.objects.get(pk=column_pk)
-        data = request.get_json(force=True, silent=True)
-        if col and data:
-            col.title = data.get('title')
-            col.color_max_cards = data.get('color_max_cards', '#FF0000')
-            col.done_column = data.get('done_column', False)
-            col.max_cards = data.get('max_cards', 9999)
+        """
+        Update a column instance
+        :param project_pk: Project that belongs
+        :param column_pk: Id of the column
+        :return: The updated resource
+        """
 
-            # check if there is another done column
-            if col.done_column:
-                columns = Column.objects(project=col.project, done_column=True)
-                for c in columns:
-                    c.done_column = False
-                    c.save()
-            col.save()
+        col = Column.get_by_id(column_pk)
+        if not col:
+            return api_errors.MissingResource(
+                api_errors.INVALID_OBJECT_ID_MSG
+            )
 
-            # save activity
-            save_notification(project_pk=col.project.pk,
-                              verb='update_column',
-                              data=col.to_dict())
+        data = request.get_json(silent=True)
 
-            return col.to_json(), 200
-        return jsonify({"error": 'Bad Request'}), 400
+        if not data:
+            return api_errors.InvalidAPIUsage(
+                api_errors.INVALID_JSON_BODY_MSG
+            )
+
+        col.title = data.get('title')
+        col.color_max_cards = data.get('color_max_cards', '#FF0000')
+        col.done_column = data.get('done_column', False)
+        col.max_cards = data.get('max_cards', 9999)
+
+        # check set other columns of the project to done_column in False
+        if col.done_column:
+            Column.clear_done_columns(project_pk)
+        col.save()
+
+        # save activity
+        save_notification(project_pk=project_pk,
+                          verb='update_column',
+                          data=col.to_dict())
+
+        return col.to_dict(), 200
 
     def delete(self, project_pk, column_pk):
-        col = Column.objects.get(pk=column_pk)
-        if col:
-            # save activity
-            save_notification(project_pk=col.project.pk,
-                              verb='delete_column',
-                              data=col.to_dict())
+        """
+        Delete a column resource
+        :param project_pk: Project that belongs
+        :param column_pk:Id of the column
+        :return: Nothing
+        """
+        col = Column.get_by_id(column_pk)
+        if not col:
+            return api_errors.MissingResource(
+                api_errors.INVALID_OBJECT_ID_MSG
+            )
+        # save activity
+        save_notification(project_pk=project_pk,
+                          verb='delete_column',
+                          data=col.to_dict())
 
-            col.delete()
-            return jsonify({"success": True}), 200
-        return jsonify({"error": 'Bad Request'}), 400
+        col.delete()
+        return {}, 204
 
 
 class ProjectColumnsOrder(AuthResource):
-    def __init__(self):
-        super(ProjectColumnsOrder, self).__init__()
+
 
     def post(self, project_pk):
         data = request.get_json(force=True, silent=True)
